Extract network from main server class
